webApp-back/requirements.txt
fastapi
uvicorn
websockets
opencv-python
numpy
psycopg2
deepface
dlib
torch
python-dotenv
alembic